+++ scripts/pf2e/Hooks.js
@@ -151,8 +151,8 @@ export function setHooks(){
     Hooks.on("preUpdateCombat",(...args) => {
        let advanceTime = args[2].advanceTime;
         if(advanceTime != 0 && !(SIFT.Settings.timeProcessor=="SimpleCalendar")) SIFT.utils.ageTemplates(advanceTime);
-        SIFT.utils.cleanupTemplates(args[0].combatant.actor.id);
-	SIFT.utils.manageUnmanaged();
+        SIFT.utils.cleanupTemplates(args[0].combatant.token.id);
+	    SIFT.utils.manageUnmanaged();
     });   
         
 }