+++ scripts/shared/Hooks.js
@@ -58,7 +58,7 @@ export function setHooks(){
         let my = e.y;
         mx += 30;
         my -= 30;
-        let ttplayer = game.actors.get(sourceTemplate.data.flags.siftoolkit?.actor)?.name??(game.users.get(sourceTemplate.data.user)?.name??"Unknown");
+        let ttplayer = game.scenes.viewed.tokens.get(sourceTemplate.data.flags.siftoolkit?.token).name??game.actors.get(sourceTemplate.data.flags.siftoolkit?.actor)?.name??(game.users.get(sourceTemplate.data.user)?.name??"Unknown");
         let ttspell = sourceTemplate.data.flags.siftoolkit?.sif??"???"
         let ttduration = "";
         
@@ -138,6 +138,24 @@ export function setHooks(){
                 button: true
             }
         );
+    });
+
+    Hooks.on("deleteCombatant", async function(...args) {        
+        let response = await SIFT.UI.promptRemoveCombatant(args);
+        let scene = args[0].combat.scene.id;
+        let token = args[0].token?.id;
+        if(response == "Remove"){
+            SIFT.utils.removeTokenTemplates(token,scene);
+        }
+    });
+
+    Hooks.on("deleteToken", async function(...args) {
+        let response = await SIFT.UI.promptRemoveCombatant(args);
+        let token = args[0].id;
+        let scene = args[0].parent?.id;
+        if(response == "Remove"){
+            SIFT.utils.removeTokenTemplates(token,scene);
+        }
     });
     
 }